backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.4.0 
torchvision==0.19.0
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.26.0
pydantic==2.5.0
transformers>=4.46.0
tokenizers>=0.20.0
huggingface_hub>=0.25.0
firebase-admin==6.2.0
openai==0.28.1